requirement.txt
Fastapi
torch
youtube-transcript-api 
nltk 
transformers
fastapi[all]
uvicorn
google-generativeai